Move MovieDetail to components/movies
