+++ src/navigations/HomeStackNavigation.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import Home from '../screens/Home'
-import MovieDetail from '../screens/MovieDetail'
+import MovieDetail from '../components/movies/MovieDetail'
 import { HomeStackParamList } from '../types/navigationTypes'
 
 const Stack = createNativeStackNavigator<HomeStackParamList>()
